+++ src/store/user/types.ts
@@ -1,0 +1,6 @@
+import { UserData } from 'api/Auth/types';
+
+export type UserState = {
+  data: UserData | null;
+  isLoading: boolean;
+};
