+++ tsconfig.json
@@ -24,6 +24,7 @@
     // "typings/**/*.d.ts",
     // "src/declarations.d.ts",
     // "./node_modules/@types",
+    "src",
     "webpack"
   ],
   "exclude": ["**/*.js", "./node_modules/"]
